+++ scripts/cleanup.sh
@@ -1,8 +1,3 @@
 #!/usr/bin/env bash
 
 set -ex -o pipefail
-
-# Setup environment
-cd `dirname $0`
-source ./env.sh
-cd ../..